+++ lib/msal-common/test/crypto/PopTokenGenerator.spec.ts
@@ -1,10 +1,8 @@
 import {
     RANDOM_TEST_GUID,
     TEST_POP_VALUES,
-    TEST_DATA_CLIENT_INFO,
     TEST_CONFIG,
     TEST_URIS,
-    TEST_CRYPTO_VALUES,
 } from "../test_kit/StringConstants.js";
 import { PopTokenGenerator } from "../../src/crypto/PopTokenGenerator.js";
 import { ICrypto } from "../../src/crypto/ICrypto.js";
@@ -14,78 +12,14 @@ import { UrlString } from "../../src/url/UrlString.js";
 import { AuthenticationScheme } from "../../src/utils/Constants.js";
 import { SignedHttpRequest } from "../../src/crypto/SignedHttpRequest.js";
 import { Logger } from "../../src/logger/Logger.js";
+import { mockCrypto } from "../client/ClientTestUtils.js";
 
 describe("PopTokenGenerator Unit Tests", () => {
     afterEach(() => {
         jest.restoreAllMocks();
     });
 
-    const cryptoInterface: ICrypto = {
-        createNewGuid(): string {
-            return RANDOM_TEST_GUID;
-        },
-        base64Decode(input: string): string {
-            switch (input) {
-                case TEST_POP_VALUES.ENCODED_REQ_CNF:
-                    return TEST_POP_VALUES.DECODED_REQ_CNF;
-                case TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO:
-                    return TEST_DATA_CLIENT_INFO.TEST_DECODED_CLIENT_INFO;
-                case TEST_POP_VALUES.SAMPLE_POP_AT_PAYLOAD_ENCODED:
-                    return TEST_POP_VALUES.SAMPLE_POP_AT_PAYLOAD_DECODED;
-                default:
-                    return input;
-            }
-        },
-        base64Encode(input: string): string {
-            switch (input) {
-                case "123-test-uid":
-                    return "MTIzLXRlc3QtdWlk";
-                case "456-test-uid":
-                    return "NDU2LXRlc3QtdWlk";
-                case TEST_POP_VALUES.DECODED_REQ_CNF:
-                    return TEST_POP_VALUES.ENCODED_REQ_CNF;
-                case TEST_POP_VALUES.SAMPLE_POP_AT_PAYLOAD_DECODED:
-                    return TEST_POP_VALUES.SAMPLE_POP_AT_PAYLOAD_ENCODED;
-                default:
-                    return input;
-            }
-        },
-        base64UrlEncode(input: string): string {
-            switch (input) {
-                case '{"kid": "XnsuAvttTPp0nn1K_YMLePLDbp7syCKhNHt7HjYHJYc"}':
-                    return "e2tpZDogIlhuc3VBdnR0VFBwMG5uMUtfWU1MZVBMRGJwN3N5Q0toTkh0N0hqWUhKWWMifQ";
-                case '{"kid":"NzbLsXh8uDCcd-6MNwXF4W_7noWXFZAfHkxZsRGC9Xs","xms_ksl":"sw"}':
-                    return "eyJraWQiOiJOemJMc1hoOHVEQ2NkLTZNTndYRjRXXzdub1dYRlpBZkhreFpzUkdDOVhzIiwieG1zX2tzbCI6InN3In0";
-                default:
-                    return input;
-            }
-        },
-        encodeKid(input: string): string {
-            switch (input) {
-                case "XnsuAvttTPp0nn1K_YMLePLDbp7syCKhNHt7HjYHJYc":
-                    return "eyJraWQiOiAiWG5zdUF2dHRUUHAwbm4xS19ZTUxlUExEYnA3c3lDS2hOSHQ3SGpZSEpZYyJ9";
-                case "NzbLsXh8uDCcd-6MNwXF4W_7noWXFZAfHkxZsRGC9Xs":
-                    return "eyJraWQiOiJOemJMc1hoOHVEQ2NkLTZNTndYRjRXXzdub1dYRlpBZkhreFpzUkdDOVhzIiwieG1zX2tzbCI6InN3In0";
-                default:
-                    return input;
-            }
-        },
-        async getPublicKeyThumbprint(): Promise<string> {
-            return TEST_POP_VALUES.KID;
-        },
-        async signJwt(): Promise<string> {
-            return "";
-        },
-        async removeTokenBindingKey(): Promise<boolean> {
-            return Promise.resolve(true);
-        },
-        async clearKeystore(): Promise<boolean> {
-            return Promise.resolve(true);
-        },
-        async hashString(): Promise<string> {
-            return Promise.resolve(TEST_CRYPTO_VALUES.TEST_SHA256_HASH);
-        },
-    };
+    const cryptoInterface: ICrypto = mockCrypto;
 
     describe("generateCnf", () => {
         const testRequest = {